+++ CHANGELOG.md
@@ -1,2 +1,4 @@
-## [](/compare/0c59a2334fbd...) (2024-05-17)
+## [](https://github.com/nao1215/honeycomb/compare/0c59a2334fbd...) (2024-05-18)
 
+* Bump github.com/google/go-cmp from 0.2.0 to 0.6.0 [#2](https://github.com/nao1215/honeycomb/pull/2) ([dependabot[bot]](https://github.com/apps/dependabot))
+* Add project template [#1](https://github.com/nao1215/honeycomb/pull/1) ([nao1215](https://github.com/nao1215))